==> 01-clean-species.ipynb <==
import pandas as pd

# Before reading in the species csv: I found an error while importing the clean dataset into SQL on line 30775
# There was an open quotation mark without the closing pair - it was causing several rows to be read as one value
# I viewed the file in a notepad and deleted the extraneous quotation mark

# read in the cvs
species_path = "../Resources/species.csv"
species_df = pd.read_csv(species_path)
species_df.head()

# drop all unnecessary columns
clean_species_df = species_df[['Park Name', 'Category', 'Common Names', 'Abundance', 'Conservation Status']]
clean_species_df

# rename columns, remove spaces to make them SQl-friendly
clean_species_df = clean_species_df.rename(columns = {"Park Name": "national_park"}).sort_values('national_park')
clean_species_df = clean_species_df.rename(columns={'Category': 'category', 'Common Names': 'animal_name', \
                                 'Abundance': 'abundance', 'Conservation Status': 'conservation_status'})

# Review
print('Describe' + '-'*100)
print(clean_species_df.describe())
print('Abundance counts' + '-'*100)
print(clean_species_df['abundance'].value_counts())
print('Category counts' + '-'*100)
print(clean_species_df['category'].value_counts())
print('Status counts' + '-'*100)
print(clean_species_df['conservation_status'].value_counts())

# Wow --- there's an extinct species on the list
extinct = clean_species_df.loc[clean_species_df['conservation_status'] == 'Extinct']
extinct

# Attempted to resolve issue with two rows. 30775-30786, which had an extraneous quotatoin mark
# and 62968, which had an extra single quote

# problem1 = clean_species_df.iloc[30775,:]
# print(problem1)
# #clean_species_df = clean_species_df.replace('"Tropical', 'Tropical', regex=True)

# problem2 = clean_species_df.iloc[62968,:]
# problem2

# Review national park names
clean_species_df['national_park'].unique()

# Save to Outputs
clean_species_df.to_csv("../Outputs/species_stats.csv", index=False)
