==> 02-clean-visitations.ipynb <==
import pandas as pd

file_path = "../Resources/park_visitation.csv"
park_visitation_df = pd.read_csv(file_path)
park_visitation_df.head()

park_visitation_df['YearRaw'].unique()

new_park_visitation = park_visitation_df.loc[(park_visitation_df['YearRaw'] == '2016') & (park_visitation_df['Unit Type'] == 'National Park')]
new_park_visitation.head()

new_park_visitation.describe()

park_visits_df = new_park_visitation[['Unit Name', 'Visitors']].copy()
park_visits_df.head()

clean_visits_df = park_visits_df.rename(columns = {"Unit Name": "national_park", "Visitors": "visitors"})
clean_visits_df = clean_visits_df.sort_values('national_park')
clean_visits_df

clean_visits_df['national_park'].unique()

visits_replace = {
    "Hawai'i": "Hawaii",
    "Wrangell-St.": "Wrangell - St",
}
clean_visits_df = clean_visits_df.replace(visits_replace, regex=True)
clean_visits_df['national_park'] = clean_visits_df['national_park'].str.strip()
clean_visits_df['national_park'].unique()

len(clean_visits_df['national_park'].unique())

# Save to Outputs
clean_visits_df.to_csv("../Resources/clean_visits.csv", index=False)
