==> 03-clean-nps.ipynb <==
#import dependencies
import pandas as pd

#read in csv
csv_path = "Resources/nps_acres.csv"
nps_acres_df = pd.read_csv(csv_path)
nps_acres_df.head()

#review park names and total of unique names
nps_unique = nps_acres_df['Park Name'].unique()
print(len(nps_unique))
print(nps_unique)

#review park names alphabetically with value counts
nps_acres_df['Park Name'].value_counts()
nps_acres_df['Park Name'].sort_values()

#Start a new data set to not change original; used Leah's code to convert abbreviations to match other datasets
clean_nps_acres = nps_acres_df.replace({
    'AL': 'Alabama',
    'AK': 'Alaska',
    'AZ': 'Arizona',
    'AR': 'Arkansas',
    'CA': 'California',
    'CO': 'Colorado',
    'CT': 'Connecticut',
    'DE': 'Delaware',
    'FL': 'Florida',
    'GA': 'Georgia',
    'HI': 'Hawaii',
    'ID': 'Idaho',
    'IL': 'Illinois',
    'IN': 'Indiana',
    'IA': 'Iowa',
    'KS': 'Kansas',
    'KY': 'Kentucky',
    'LA': 'Louisiana',
    'ME': 'Maine',
    'MD': 'Maryland',
    'MA': 'Massachusetts',
    'MI': 'Michigan',
    'MN': 'Minnesota',
    'MS': 'Mississippi',
    'MO': 'Missouri',
    'MT': 'Montana',
    'NE': 'Nebraska',
    'NV': 'Nevada',
    'NH': 'New Hampshire',
    'NJ': 'New Jersey',
    'NM': 'New Mexico',
    'NY': 'New York',
    'NC': 'North Carolina',
    'ND': 'North Dakota',
    'OH': 'Ohio',
    'OK': 'Oklahoma',
    'OR': 'Oregon',
    'PA': 'Pennsylvania',
    'RI': 'Rhode Island',
    'SC': 'South Carolina',
    'SD': 'South Dakota',
    'TN': 'Tennessee',
    'TX': 'Texas',
    'UT': 'Utah',
    'VT': 'Vermont',
    'VA': 'Virginia',
    'WA': 'Washington',
    'WV': 'West Virginia',
    'WI': 'Wisconsin',
    'WY': 'Wyoming'
})

#review state names, noticing there may be issues with joining with parks identified in multiple states 
clean_nps_acres['state'].value_counts()

#rename columns so they can be read in sql wihtout spaces and captials
clean_nps_acres = clean_nps_acres.rename(columns={"Park Code": "park_code", "Park Name": "national_park", "State": "state", "Acres": "acres", "Latitude": 'latitude', "Longitude":"longitude"})
clean_nps_acres.head(1)

#Leah's cleaning code added
nps_replace = {
    "and Preserve": "",
    "The ": "the ",
    "Of ": "of "
}
clean_nps_acres = clean_nps_acres.replace(nps_replace, regex=True)
clean_nps_acres['national_park'] = clean_nps_acres['national_park'].str.strip()
clean_nps_acres['national_park'].unique()

nps_df = pd.DataFrame(clean_nps_acres)
nps_df.head()

#save original data as csv
nps_df.to_csv('Resources/nps_df.csv', index=False, header=True)
