==> 04-clean-alltrails.ipynb <==
#import dependencies
import pandas as pd

#read in dataset
csv_path = "Resources/AllTrails_data_nationalpark.csv"
all_trails_df = pd.read_csv(csv_path)
all_trails_df.head()

#review visitor_usage for min and max
all_trails_df['visitor_usage'].describe()
print(visitor_usage)

#review trail name value counts and length
all_trails_df['name'].value_counts()

#review country name since this is supposed to be a US National Park dataset 
all_trails_df['country_name'].value_counts()

#reive state names with value counts
all_trails_df['state_name'].value_counts()

#review park names and unique length
area_unique = all_trails_df['area_name'].unique()
print(len(area_unique))
area_unique

#review features with value_counts and length
all_trails_df['features'].value_counts()

#review activities with value_counts and length; notice difference between 744 for features and 526 for activities
all_trails_df['activities'].value_counts()

#after discussing the review of this dataset; agree to drop specific columns
clean_all_trails = all_trails_df.drop(columns=['city_name', 'country_name', 'units', 'visitor_usage', 'features', 'activities'])
clean_all_trails.head()

#After further review of national parks, google results inform me that this Clayton Co International is not a national park.
print(clean_all_trails[clean_all_trails['area_name']=="Clayton Co International Park, Jonesboro GA"])

#dropping row with Clayton Co International Park by index
clean_all_trails.drop([1219], inplace = True)

#Leah found this code to help split the _geoloc column
lat = []
lon = []
# For each row in a varible,
for row in clean_all_trails['_geoloc']:
    # Try to,
    try:
        # Split the row by comma and append
        # everything before the comma to lat
        lat.append(row.split(',')[0])
        # Split the row by comma and append
        # everything after the comma to lon
        lon.append(row.split(',')[1])
    # But if you get an error
    except:
        # append a missing value to lat
        lat.append(np.NaN)
        # append a missing value to lon
        lon.append(np.NaN)
# Create two new columns from lat and lon
clean_all_trails['trailhead_lat'] = lat
clean_all_trails['trailhead_long'] = lon

#dropping _geoloc now that we have trailhead_lat and trailhead_long
clean_all_trails = clean_all_trails.drop(columns=['_geoloc'])
clean_all_trails.head()

#dataframe before next step
clean_trails_df = pd.DataFrame(clean_all_trails)
clean_trails_df.head()

#Leah's help with further cleaning these columns
cleaning_lat = {
    "{'lat':": ""
}

clean_trails_df = clean_trails_df.replace(cleaning_lat, regex=True)
clean_trails_df['trailhead_lat'] = clean_trails_df['trailhead_lat'].str.strip()

clean_trails_df.head()

cleaning_lon = {
    "'lng':": "",
    "}": ""
}
clean_trails_df = clean_trails_df.replace(cleaning_lon, regex=True)
clean_trails_df['trailhead_long'] = clean_trails_df['trailhead_long'].str.strip()

clean_trails_df.head()

#renaming columns to match schema
clean_trails_df = clean_trails_df.rename(columns={"name": "trail_name",
                                                    "area_name": "national_park",
                                                    "state_name": "state",
                                                    "elevation_gain": "elevation",
                                                    "difficulty_rating": "difficulty",
                                                    "route_type": "route_type",
                                                    "avg_rating": "rating",
                                                   })
clean_trails_df.head()

#rounding popularity
clean_trails_df['popularity'] = clean_trails_df['popularity'].round(1)
popularity_stats = clean_trails_df['popularity'].describe()
print(popularity_stats)
clean_trails_df.head()

#rounding elevation
clean_trails_df['elevation'] = clean_trails_df['elevation'].round(1)
elevation_stats = clean_trails_df['elevation'].describe()
print(elevation_stats)
clean_trails_df.head()

#reviewing the length column
clean_trails_df['length'].describe()

#moving decimal to make values in kilometers
clean_trails_df['length'] = clean_trails_df['length'].div(1000)
clean_trails_df.head()

#changing length value of kilometers to miles
clean_trails_df['length'] = clean_trails_df['length'].div(1.609)
clean_trails_df.head()

#rounding length
clean_trails_df['length'] = clean_trails_df['length'].round(1)
length_stats = clean_trails_df['length'].describe()
print(length_stats)
clean_trails_df.head()

clean_trails_df.to_csv('Resources/clean_trails_df.csv', index=False, header=True)
