==> 05-merge-dataframe.ipynb <==
import pandas as pd

csv_path = "../Resources/clean_trails_df.csv"
clean_trails_df = pd.read_csv(csv_path)
clean_trails_df.head()

csv_path = "../Resources/nps_df.csv"
nps_df = pd.read_csv(csv_path)
nps_df.head()

nps_park_diff = set(nps_df.national_park).difference(clean_trails_df.national_park)
print(nps_park_diff)

trails_park_diff = set(clean_trails_df.national_park).difference(nps_df.national_park)
print(trails_park_diff)

merge_clean_trails = pd.DataFrame.merge(nps_df, clean_trails_df, how='left')
merge_clean_trails.head()

unique_parks = merge_clean_trails['national_park'].unique()
unique_parks
len(unique_parks)

unique_parks

merge_clean_trails.info()

merge_clean_trails['state'].value_counts()

trail_stats = merge_clean_trails.drop(columns=['national_park', 'state', 'acres', 'latitude', 'longitude', 'trail_name', 'trailhead_lat', 'trailhead_long'])
trail_stats.head()

trail_stats.to_csv('Output/trail_stats.csv', index=False, header=True)

trails_df = clean_trails_df.drop(columns=['popularity', 'length', 'elevation', 'difficulty', 'route_type', 'rating', 'num_reviews'])
trails_df.head()

trails_df.to_csv('Output/trails.csv', index=False, header=True)

csv_path = "Resources/nps_df.csv"
clean_nps_df = pd.read_csv(csv_path)
clean_nps_df.head()

visits_path = "Resources/clean_visits.csv"
clean_visits_df = pd.read_csv(visits_path)
clean_visits_df.head()

merge_nps_df = pd.DataFrame.merge(clean_nps_df, clean_visits_df, how="left")
merge_nps_df.head()

merge_nps_df['national_park'].value_counts()

merge_nps_df.count()

merge_nps_df.to_csv("../Outputs/national_parks.csv", index=False)

species_path = "../Outputs/species_stats.csv"
species_df = pd.read_csv(species_path)
species_df.head()

merge_species = pd.DataFrame.merge(merge_nps_df, species_df, how="left")
merge_species = merge_species.drop(columns = ['state', 'acres', 'latitude', 'longitude', 'visitors'])
merge_species

merge_species.to_csv("../Outputs/species_stats.csv", index=False)
